+++ src/main/kotlin/de/achimonline/kickassembler/acbg/settings/KickAssemblerSettings.kt
@@ -1,0 +1,5 @@
+package de.achimonline.kickassembler.acbg.settings
+
+data class KickAssemblerSettings(
+    var jreNameOrPath: String? = null
+)